+++ scripts/package.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line no-console, @typescript-eslint/no-var-requires
-console.log(require("../package.json")[process.argv[2]]);