+++ src/WithOptionalObjectId.ts
@@ -1,5 +1,5 @@
-import type { ObjectId } from "mongodb";
+import type { Document, ObjectId } from "mongodb";
 
-export interface WithOptionalObjectId {
+export interface WithOptionalObjectId extends Document {
   _id?: ObjectId;
 }